requirements.txt
torch==2.3.0
transformers==4.44.0
datasets==2.20.0
peft==0.12.0
trl==0.9.6
accelerate==0.33.0
bitsandbytes==0.43.3
vllm==0.5.4
sentence-transformers==3.0.1
chromadb==0.5.3
sqlglot==23.12.2
openai==1.40.0
wandb==0.17.6
rich==13.7.1
typer==0.12.3
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
fastapi==0.112.0
uvicorn==0.30.5
einops==0.8.0
scipy==1.14.0
scikit-learn==1.5.1
pyyaml>=6.0
